popApp addCat, create categories list
