+++ client/src/components/Categories/AddCategories.module.css
@@ -1,0 +1,76 @@
+
+.gral{
+  margin-top: 80px;
+
+}
+
+
+.form {
+    width: 450px;
+    padding: 40px;
+    margin: auto;
+    border-radius: 20px;
+    border-color: black;
+    color: white;
+    font-size: 16px;
+    background-color: var(--color-principal-violet);
+    border-radius: 20px;
+    -webkit-box-shadow: 0px 9px 13px -2px #9e9e9e;
+    box-shadow: 0px 9px 13px -2px #969696;
+    /* background-position-x: -185px;
+    background-position-y: -50px; */
+    font-weight: 700;
+    margin-top: 55px;
+  }
+
+  .form input {
+    width: 100%;
+    padding: 10px;
+    border-radius: .25rem;
+    margin-top: 20px;
+    margin-bottom: 20px;
+    border-style: none;
+    /* border-color: gray;    */
+  }
+
+  .btn {
+    border-style: solid;
+    border-color: transparent;
+    border-radius: .25rem;
+    cursor: pointer;
+    margin: 2rem;
+    padding: 8px 25px 8px 25px;
+    background-color: var(--color-pink-buttons);
+    color: white;
+  }
+  .btn:hover {
+    background-color: var(--color-pink-graphics);
+    font-weight: bold;
+    color: var(--color-principal-violet);
+  }
+
+  .select {
+    width: 70%;
+    padding: 1.5%;
+    border-radius: .25rem;
+    color: grey;
+    font-size: 14px;
+  }
+
+  .close {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+  
+    width: 1.5rem;
+    height: 1.5rem;
+    border-radius: .25rem;
+    margin-top: -5%;
+    margin-right: -5%;
+    cursor: pointer;
+  }
+  
+  .close:hover {
+    font-weight: 700;
+    background-color: var(--color-violet);
+  }
