+++ client/src/components/Categories/Categories.module.css
@@ -1,58 +1,21 @@
-
-.gral{
-  margin-top: 80px;
-
+.addCat {
+    padding: 1rem;
+    border-radius: .25rem;
+    border-color: transparent;
+    background-color: var(--color-principal-violet);
+    color: white;
+    margin-bottom: 2rem;
+    cursor: pointer;
 }
 
-
-.form {
-    width: 450px;
-    padding: 40px;
-    margin: auto;
-    border-radius: 20px;
-    border-color: black;
-    color: white;
-    font-size: 16px;
-    background-color: var(--color-principal-violet);
-    border-radius: 20px;
-    -webkit-box-shadow: 0px 9px 13px -2px #9e9e9e;
-    box-shadow: 0px 9px 13px -2px #969696;
-    /* background-position-x: -185px;
-    background-position-y: -50px; */
-    font-weight: 700;
-    margin-top: 55px;
-  }
-
-  .form input {
-    width: 100%;
-    padding: 10px;
+.btnCat {
+    padding: 1rem;
+    width: 15rem;
     border-radius: .25rem;
-    margin-top: 20px;
-    margin-bottom: 20px;
-    border-style: none;
-    /* border-color: gray;    */
-  }
-
-  .btn {
-    border-style: solid;
-    border-color: transparent;
-    border-radius: .25rem;
-    cursor: pointer;
-    margin: 2rem;
-    padding: 8px 25px 8px 25px;
-    background-color: var(--color-pink-buttons);
-    color: white;
-  }
-  .btn:hover {
-    background-color: var(--color-pink-graphics);
-    font-weight: bold;
+    border-width: .1rem;
+    background-color: var(--color-light-violet);
     color: var(--color-principal-violet);
-  }
-
-  .select {
-    width: 70%;
-    padding: 1.5%;
-    border-radius: .25rem;
-    color: grey;
-    font-size: 14px;
-  }
+    font-weight: 700;
+    cursor: pointer;
+    margin: .75rem;
+}
